requirements.txt
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
scikit-learn==1.4.1.post1
shap==0.45.0
PyYAML==6.0.1
